move train and eval anyd functions to a new joint helper file
